+++ src/WebJobs.Script/Extensions/FunctionMetadataExtensions.cs
@@ -52,14 +52,23 @@ namespace Microsoft.Azure.WebJobs.Script
         public static bool IsProxy(this FunctionMetadata metadata) =>
             metadata is ProxyFunctionMetadata;
 
-        public static bool IsDirect(this FunctionMetadata metadata) =>
-            GetBoolProperty(metadata.Properties, IsDirectKey);
+        public static bool IsDirect(this FunctionMetadata metadata)
+        {
+            BoolUtility.TryReadAsBool(metadata.Properties, IsDirectKey, out bool result);
+            return result;
+        }
 
-        public static bool IsDisabled(this FunctionMetadata metadata) =>
-            GetBoolProperty(metadata.Properties, IsDisabledKey);
+        public static bool IsDisabled(this FunctionMetadata metadata)
+        {
+            BoolUtility.TryReadAsBool(metadata.Properties, IsDisabledKey, out bool result);
+            return result;
+        }
 
-        public static bool IsCodeless(this FunctionMetadata metadata) =>
-            GetBoolProperty(metadata.Properties, IsCodelessKey);
+        public static bool IsCodeless(this FunctionMetadata metadata)
+        {
+            BoolUtility.TryReadAsBool(metadata.Properties, IsCodelessKey, out bool result);
+            return result;
+        }
 
         public static bool IsCodelessSet(this FunctionMetadata metadata) =>
             metadata.Properties.ContainsKey(IsCodelessKey);
@@ -88,12 +97,5 @@ namespace Microsoft.Azure.WebJobs.Script
         /// </summary>
         public static void SetIsCodeless(this FunctionMetadata metadata, bool value) =>
             metadata.Properties[IsCodelessKey] = value;
-
-        private static bool GetBoolProperty(IDictionary<string, object> properties, string propertyKey)
-        {
-            return properties.TryGetValue(propertyKey, out object valObj)
-                && valObj is bool valBool
-                && valBool;
-        }
     }
 }