+++ release_notes.md
@@ -1,4 +1,5 @@
 ### Release notes
+
 <!-- Please add your release notes in the following format:
 - My change description (#PR)
 -->
@@ -6,6 +7,7 @@
 - Update Java Worker Version to [2.7.0](https://github.com/Azure/azure-functions-java-worker/releases/tag/2.7.0)
 - Update Node.js Worker Version to [3.5.1](https://github.com/Azure/azure-functions-nodejs-worker/releases/tag/v3.5.1)
 - Update ApplicationInsights packages to 2.21.0 (#8838)
+- Support model-type bindings for out-of-proc workers (#8815)
 
 **Release sprint:** Sprint 131
 [ [bugs](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+131%22+label%3Abug+is%3Aclosed) | [features](https://github.com/Azure/azure-functions-host/issues?q=is%3Aissue+milestone%3A%22Functions+Sprint+131%22+label%3Afeature+is%3Aclosed) ]